backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
websockets==13.0
stripe==10.0.0
chromadb==0.5.0
aiofiles==24.1.0
python-dotenv==1.0.1
